requirements.txt
Pillow=8.1.0
django-richtextfield=1.6
django-ckeditor=6.0.0